Edges and VertexCollection as a separate interface GraphCollection
